Use CASTLE_FLAG_BANK to merge e_misc
